backend/requirements.txt
Flask==3.0.3
pyserial==3.5.0
numpy==1.26.4
scipy==1.13.0
torch==2.2.2
websocket-client==1.7.0
